remove unnecessary ASCII filtering
