+++ src/hasher.rs
@@ -22,6 +22,12 @@ pub trait Hasher: Sync + Send {
   /// The hash algorithm identifier MUST be a hash algorithm value from the
   /// "Hash Name String" column in the IANA "Named Information Hash Algorithm"  
   fn alg_name(&self) -> &'static str;
+
+  /// Returns the base64url-encoded digest of a `disclosure`.
+  fn encoded_digest(&self, disclosure: &str) -> String {
+    let hash = self.digest(disclosure.as_bytes());
+    multibase::Base::Base64Url.encode(hash)
+  }
 }
 
 /// An implementation of [`Hasher`] that uses the `sha-256` hash function.
@@ -45,5 +51,36 @@ impl Hasher for Sha256Hasher {
 
   fn alg_name(&self) -> &'static str {
     Sha256Hasher::ALG_NAME
+  }
+}
+
+// Some test values taken from https://www.ietf.org/archive/id/draft-ietf-oauth-selective-disclosure-jwt-05.html#name-hashing-disclosures
+#[cfg(test)]
+mod test {
+  use crate::{Hasher, Sha256Hasher};
+
+  #[test]
+  fn test1() {
+    let disclosure = "WyI2cU1RdlJMNWhhaiIsICJmYW1pbHlfbmFtZSIsICJNw7ZiaXVzIl0";
+    let hasher = Sha256Hasher::new();
+    let hash = hasher.encoded_digest(disclosure);
+    assert_eq!("uutlBuYeMDyjLLTpf6Jxi7yNkEF35jdyWMn9U7b_RYY", hash);
+  }
+
+  #[test]
+  fn test2() {
+    let disclosure =
+      "WyJlSThaV205UW5LUHBOUGVOZW5IZGhRIiwgImVtYWlsIiwgIlwidW51c3VhbCBlbWFpbCBhZGRyZXNzXCJAZXhhbXBsZS5qcCJd";
+    let hasher = Sha256Hasher::new();
+    let hash = hasher.encoded_digest(disclosure);
+    assert_eq!("Kuet1yAa0HIQvYnOVd59hcViO9Ug6J2kSfqYRBeowvE", hash);
+  }
+
+  #[test]
+  fn test3() {
+    let disclosure = "WyJsa2x4RjVqTVlsR1RQVW92TU5JdkNBIiwgIkZSIl0";
+    let hasher = Sha256Hasher::new();
+    let hash = hasher.encoded_digest(disclosure);
+    assert_eq!("w0I8EKcdCtUPkGCNUrfwVp2xEgNjtoIDlOxc9-PlOhs", hash);
   }
 }
