+++ src/lib.rs
@@ -10,7 +10,6 @@ mod error;
 mod hasher;
 mod key_binding_jwt_claims;
 mod sd_jwt;
-mod utils;
 
 pub use decoder::*;
 pub use disclosure::*;
@@ -22,4 +21,3 @@ pub use sd_jwt::*;
 pub use serde_json::json;
 pub use serde_json::Map;
 pub use serde_json::Value;
-pub use utils::*;
